// frontend/src/app/apiClient/model/credentialReset.ts
/**
 * LoopBack Application
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */






export interface CredentialReset { 

    
    credentialResetID?: string;

    
    email: string;

    
    newEmail?: string;

    
    newPassword?: string;


    newPhoneNumber?: string;
    
    
    userID?: string;

    
    timestamp?: string;

}