// frontend/src/app/apiClient/model/companyUser.ts
/**
 * LoopBack Application
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */



import { BigDecimal } from 'bigdecimal';
import { User, Company } from '..';




export interface CompanyUser { 

    
    companyUserID: string;

    
    companyID: string;

    
    userID: BigDecimal;

    
    lastLogin: Date;

    user?: User;

    company?: Company;

}